authorization implemented

+++ src/api/notes/handler.js
@@ -13,8 +13,14 @@ class NotesHandler {
 	async postNoteHandler(request, h) {
 		this._validator.validateNotePayload(request.payload);
 		const { title = "untitled", body, tags } = request.payload;
+		const { id: credentialsId } = request.auth.credentials;
 
-		const noteId = await this._service.addNote({ title, body, tags });
+		const noteId = await this._service.addNote({
+			title,
+			body,
+			tags,
+			owner: credentialsId,
+		});
 
 		const response = h.response({
 			status: "success",
@@ -27,8 +33,10 @@ class NotesHandler {
 		return response;
 	}
 
-	async getNotesHandler() {
-		const notes = await this._service.getNotes();
+	async getNotesHandler(request) {
+		const { id: credentialsId } = request.auth.credentials;
+
+		const notes = await this._service.getNotes(credentialsId);
 		return {
 			status: "success",
 			data: {
@@ -39,6 +47,9 @@ class NotesHandler {
 
 	async getNoteByIdHandler(request) {
 		const { id } = request.params;
+		const { id: credentialsId } = request.auth.credentials;
+
+		await this._service.verifyNoteOwner(id, credentialsId);
 		const note = await this._service.getNoteById(id);
 		return {
 			status: "success",
@@ -51,8 +62,10 @@ class NotesHandler {
 	async putNoteByIdHandler(request) {
 		this._validator.validateNotePayload(request.payload);
 		const { id } = request.params;
+		const { id: credentialsId } = request.auth.credentials;
 
-		await this._service.editNoteById(id, request.payload);
+		await this._service.verifyNoteOwner(id, credentialsId);
+		await this._service.updateNoteById(id, request.payload);
 
 		return {
 			status: "success",
@@ -62,6 +75,9 @@ class NotesHandler {
 
 	async deleteNoteByIdHandler(request) {
 		const { id } = request.params;
+		const { id: credentialsId } = request.auth.credentials;
+
+		await this._service.verifyNoteOwner(id, credentialsId);
 		await this._service.deleteNoteById(id);
 
 		return {
